Moved SpreadWrapper for use from other classes.
Wrapped with delegate to avoid exposing the SpreadWrapper.
